+++ pkg/detectors/boxoauth/boxoauth_test.go
@@ -1,6 +1,3 @@
-//go:build detectors
-// +build detectors
-
 package boxoauth
 
 import (
